Refactor - move mobile topbar code close to non-mobile version
